+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/compilation_tests/AllGraphQLCasesServer_util_batchMapping_Collection_Test.java
@@ -1,0 +1,10 @@
+package com.graphql_java_generator.plugin.compilation_tests;
+
+import graphql.mavenplugin_notscannedbyspring.AllGraphQLCases_Server_SpringConfiguration_util_batchMapping_Collection;
+
+class AllGraphQLCasesServer_util_batchMapping_Collection_Test extends AllGraphQLCasesServer_util_batchMapping_Test {
+
+	public AllGraphQLCasesServer_util_batchMapping_Collection_Test() {
+		super(AllGraphQLCases_Server_SpringConfiguration_util_batchMapping_Collection.class);
+	}
+}