+++ graphql-maven-plugin-logic/src/test/java/graphql/mavenplugin_notscannedbyspring/AllGraphQLCases_Server_SpringConfiguration_util_batchMapping_MonoMap.java
@@ -1,0 +1,34 @@
+/**
+ * 
+ */
+package graphql.mavenplugin_notscannedbyspring;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.ComponentScan.Filter;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.FilterType;
+
+import com.graphql_java_generator.plugin.conf.BatchMappingDataFetcherReturnType;
+import com.graphql_java_generator.plugin.test.helper.GraphQLConfigurationTestHelper;
+
+/**
+ * The Spring configuration used for JUnit tests
+ * 
+ * @author etienne-sf
+ */
+@Configuration
+@ComponentScan(basePackages = "com.graphql_java_generator", excludeFilters = {
+		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateRelaySchema.*"),
+		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateGraphQLSchema.*"),
+		@Filter(type = FilterType.REGEX, pattern = "com.graphql_java_generator.client.graphqlrepository.*") })
+public class AllGraphQLCases_Server_SpringConfiguration_util_batchMapping_MonoMap
+		extends AllGraphQLCases_Server_SpringConfiguration_util_batchMapping_Flux {
+
+	@Override
+	protected void addSpecificConfigurationParameterValue(GraphQLConfigurationTestHelper configuration) {
+		// We just override some parameters, so let's start with our parent's values
+		super.addSpecificConfigurationParameterValue(configuration);
+
+		configuration.batchMappingDataFetcherReturnType = BatchMappingDataFetcherReturnType.MONO_MAP;
+	}
+}