+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-Forum-server/src/main/java/org/forum/server/specific_code/DataFetchersDelegateBoardImpl.java
@@ -3,9 +3,12 @@
  */
 package org.forum.server.specific_code;
 
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 import org.dataloader.BatchLoaderEnvironment;
+import org.dataloader.DataLoader;
 import org.forum.server.graphql.Board;
 import org.forum.server.graphql.Topic;
 import org.forum.server.graphql.util.DataFetchersDelegateBoard;
@@ -16,6 +19,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import graphql.GraphQLContext;
+import graphql.schema.DataFetchingEnvironment;
 import jakarta.annotation.Resource;
 import reactor.core.publisher.Flux;
 
@@ -36,33 +40,29 @@ public class DataFetchersDelegateBoardImpl implements DataFetchersDelegateBoard 
 	TopicRepository topicRepository;
 	@Resource
 	BoardRepository boardRepository;
+	
+	public List<Topic> topics(DataFetchingEnvironment dataFetchingEnvironment, Board source, Date since) {
+		if (since == null)
+			return this.topicRepository.findByBoardId(source.getId());
+		else
+			return this.topicRepository.findByBoardIdAndSince(source.getId(), since);
+	}
 
 	@Override
-	public Flux<List<Topic>> topics(//
-			BatchLoaderEnvironment batchLoaderEnvironment, //
-			GraphQLContext graphQLContext, //
-			List<Board> boards //
-	// @Argument("since") java.util.Date since//
-	) {
-		java.util.Date since = null;
-		return Flux.fromIterable(boards).map(b -> {
-			if (since == null)
-				return this.topicRepository.findByBoardId(b.getId());
-			else
-				return this.topicRepository.findByBoardIdAndSince(b.getId(), since);
-		});
+	public Object topics(
+			DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<java.lang.Long, org.forum.server.graphql.Topic> dataLoader,
+			org.forum.server.graphql.Board origin,
+			java.util.Date since) {
+		// When the data is modeled this way (that is: in a relational database), using Data Loader is not an
+		// optimization.
+		// But this is used here for integration tests
+		List<Long> ids = new ArrayList<>();
+		for (Topic topic : topics(dataFetchingEnvironment, origin, since)) {
+			ids.add(topic.getId());
+		}
+		return dataLoader.loadMany(ids);
 
-		// Mono.create(callback -> {
-		// this.logger.debug("Batch loading {} topics", boards.size());
-		// Map<Board, List<Topic>> map = new HashMap<>();
-		// for (Board b : boards) {
-		// if (since == null)
-		// map.put(b, this.topicRepository.findByBoardId(b.getId()));
-		// else
-		// map.put(b, this.topicRepository.findByBoardIdAndSince(b.getId(), since));
-		// }
-		// callback.success(map);
-		// });
 	}
 
 }