+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-Forum-server/src/main/java/org/forum/server/specific_code/DataFetchersDelegatePostImpl.java
@@ -4,10 +4,9 @@
 package org.forum.server.specific_code;
 
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
 
 import org.dataloader.BatchLoaderEnvironment;
-import org.dataloader.DataLoader;
 import org.forum.server.graphql.Member;
 import org.forum.server.graphql.Post;
 import org.forum.server.graphql.util.DataFetchersDelegatePost;
@@ -17,8 +16,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
-import graphql.schema.DataFetchingEnvironment;
+import graphql.GraphQLContext;
 import jakarta.annotation.Resource;
+import reactor.core.publisher.Flux;
 
 /**
  * @author etienne-sf
@@ -35,9 +35,24 @@ public class DataFetchersDelegatePostImpl implements DataFetchersDelegatePost {
 	PostRepository postRepository;
 
 	@Override
-	public CompletableFuture<Member> author(DataFetchingEnvironment dataFetchingEnvironment,
-			DataLoader<Long, Member> dataLoader, Post origin) {
-		return dataLoader.load(origin.getAuthorId());
+	public Flux<Member> author(BatchLoaderEnvironment batchLoaderEnvironment, GraphQLContext graphQLContext,
+			List<Post> keys) {
+		if (this.logger.isDebugEnabled()) {
+			List<String> idList = keys.stream().map(p -> p.getAuthorId().toString()).collect(Collectors.toList());
+			this.logger.debug("Before returning Flux to load this list of posts authors: {}", String.join(",", idList));
+		}
+		return Flux.fromIterable(keys).map(obj -> {
+			this.logger.debug("Before loading Member {} ", obj.getAuthorId());
+			Member ret = this.memberRepository.findById(obj.getAuthorId()).orElse(null);
+
+			// To check that this member is loaded from the @BatchMapping controller method, we prefix the member's name
+			if (!ret.getName().startsWith("[BM] ")) {
+				ret.setName("[BM] " + ret.getName());
+			}
+
+			this.logger.debug("After loading Member {}: ", obj.getAuthorId(), ret);
+			return ret;
+		});
 	}
 
 	@Override