+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateBatchMappingTypeImpl.java
@@ -1,0 +1,18 @@
+package org.allGraphQLCases.server.impl;
+
+import org.allGraphQLCases.server.DataFetchersDelegateBatchMappingType;
+import org.allGraphQLCases.server.STP_BatchMappingType_STS;
+import org.springframework.stereotype.Component;
+
+import graphql.schema.DataFetchingEnvironment;
+
+@Component
+public class DataFetchersDelegateBatchMappingTypeImpl implements DataFetchersDelegateBatchMappingType {
+
+	@Override
+	public Object friends(DataFetchingEnvironment dataFetchingEnvironment, STP_BatchMappingType_STS origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+}